Move configuration classes to `DNS_BLM.Domain.Configuration` and update references
