--- src/DNS-BLM.Domain/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace DNS_BLM.Domain;
-
-public class Class1
-{
-}
