+++ src/DNS-BLM.Domain/DNS-BLM.Domain.csproj
@@ -6,11 +6,4 @@
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
     </PropertyGroup>
-
-    <ItemGroup>
-      <Reference Include="Microsoft.Extensions.Configuration.Abstractions">
-        <HintPath>..\..\..\..\.nuget\packages\microsoft.extensions.configuration.abstractions\9.0.3\lib\net9.0\Microsoft.Extensions.Configuration.Abstractions.dll</HintPath>
-      </Reference>
-    </ItemGroup>
-
 </Project>
